requirements.txt
django-toolbelt
python-gcm